requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
openmeteo-requests>=1.1.0
requests-cache>=1.1.0
retry-requests>=2.0.0
pandas>=2.1.1
numpy>=1.26.0
urllib3>=2.0.7
requests>=2.31.0
